+++ trunk/xorp/bgp/test_fanout.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/bgp/test_fanout.cc,v 1.3 2002/12/16 04:05:14 mjh Exp $"
+#ident "$XORP: xorp/bgp/test_fanout.cc,v 1.4 2002/12/17 22:06:07 mjh Exp $"
 
 #include "bgp_module.h"
 #include "config.h"
@@ -83,21 +83,21 @@ int main(int, char** argv) {
     OriginAttribute igp_origin_att(IGP);
 
     AsPath aspath1;
-    aspath1.add_AS_in_sequence(AsNum((uint16_t)1));
-    aspath1.add_AS_in_sequence(AsNum((uint16_t)2));
-    aspath1.add_AS_in_sequence(AsNum((uint16_t)3));
+    aspath1.add_AS_in_sequence(AsNum(1));
+    aspath1.add_AS_in_sequence(AsNum(2));
+    aspath1.add_AS_in_sequence(AsNum(3));
     ASPathAttribute aspathatt1(aspath1);
 
     AsPath aspath2;
-    aspath2.add_AS_in_sequence(AsNum((uint16_t)4));
-    aspath2.add_AS_in_sequence(AsNum((uint16_t)5));
-    aspath2.add_AS_in_sequence(AsNum((uint16_t)6));
+    aspath2.add_AS_in_sequence(AsNum(4));
+    aspath2.add_AS_in_sequence(AsNum(5));
+    aspath2.add_AS_in_sequence(AsNum(6));
     ASPathAttribute aspathatt2(aspath2);
 
     AsPath aspath3;
-    aspath3.add_AS_in_sequence(AsNum((uint16_t)7));
-    aspath3.add_AS_in_sequence(AsNum((uint16_t)8));
-    aspath3.add_AS_in_sequence(AsNum((uint16_t)9));
+    aspath3.add_AS_in_sequence(AsNum(7));
+    aspath3.add_AS_in_sequence(AsNum(8));
+    aspath3.add_AS_in_sequence(AsNum(9));
     ASPathAttribute aspathatt3(aspath3);
 
     PathAttributeList<IPv4>* palist1 =
